Feat/generalize project
* generalize labels

* rename toml config file

* unused import

* generalize show landmarks and labels cli tool

* generalize project name on wandb

* generalize tests and default dataset name

* move playground code to src

* update readme

+++ play-video.py
@@ -1,16 +1,30 @@
-import cv2
-from os import listdir
-from sys import argv
-from os.path import join
+from argparse import ArgumentParser
+from os.path import exists
 
-from src.video.play.with_text import play_with_frame_num_detailed
+if __name__ == "__main__":
 
-if __name__ == '__main__':
-    videos_path = "./data/videos"
-    videos = listdir(videos_path)
-    video_idx = int(argv[1])
+    parser = ArgumentParser(
+        prog="play-video",
+        description=f"""
+            Play the given video with detailed controls and annotated with the current frame number.
+            Controls during playback: 
+            Spacebar: pause/play
+            left arrow key: rewind video. 1s while playing, 1 frame while paused.
+            right arrow key: forward video. 1s while playing, 1 frame while paused. 
+            q key: quit video playback.
+        """,
+    )
+    parser.add_argument("path", help="Path to video to play.")
+    args = parser.parse_args()
+    video_name = args.path
+    if not exists(video_name):
+        raise Exception(f"Video file {video_name} not found, cannot play.")
+
+    import cv2
+
+    from src.video.play.with_text import play_with_frame_num_detailed
 
     try:
-        play_with_frame_num_detailed(join(videos_path, videos[video_idx]))
+        play_with_frame_num_detailed(video_name)
     finally:
         cv2.destroyAllWindows()
